backend/uta-processor/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
camelot-py==1.0.0
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
et_xmlfile==2.0.0
fastapi==0.115.12
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
idna==3.10
numpy==2.2.4
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pdfminer.six==20250327
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pypdf==5.4.0
pypdfium2==4.30.1
python-dateutil==2.9.0.post0
python-magic==0.4.27
python-multipart==0.0.20
pytz==2025.1
requests==2.32.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.39
starlette==0.46.1
tabulate==0.9.0
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.34.0
